+++ .gitignore
@@ -12,6 +12,7 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 pnpm-debug.log*
+.vite-ssg-temp
 
 # Editor directories and files
 .idea
