+++ src/components/Islands/Island.vue
@@ -63,6 +63,7 @@ export default defineComponent({
         </div>
         <island-title :checked="island.status >= 2" v-if="island.status !== 0" :color="islandInfos.color" small :name="island.name">{{ island.name }}</island-title>
       </div>
+
     </foreignObject>
   </g>
 </template>
