+++ src/main.ts
@@ -3,17 +3,14 @@ import { ViteSSG } from 'vite-ssg'
 
 import App from './App.vue'
 import generatedRoutes from '~pages'
-import './registerServiceWorker'
 import './assets/scss/main.scss'
 import type { UserModule } from '@/types'
-
 const routes = setupLayouts(generatedRoutes)
 
 export const createApp = ViteSSG(
   App,
-  { routes, base: '/coucou' },
+  { routes, base: '/' },
   (ctx) => {
-    // install all modules under `modules/`
     Object.values(import.meta.glob<{ install: UserModule }>('./modules/*.ts', { eager: true }))
       .forEach(i => i.install?.(ctx))
   },
